yolo_project/requirements.txt
ultralytics>=8.0.0
pdf2image
opencv-python
Pillow
numpy
tqdm